Ef search feature

+++ player-search.php
@@ -6,10 +6,6 @@ include './database/db-helpers.php';
 
 loginCheck();
 
-$sql = "SELECT RLPID, name, position, mvps, goals, assists FROM RLPlayer";
-
-$data = execute_query($sql);
-
 $teamid = execute_query("SELECT TeamID FROM Team WHERE UserID=?", array($_SESSION['UserID']))['rows_affected'][0]['TeamID'];
 
 // Current Team Members
@@ -20,34 +16,51 @@ for ($i = 0; $i < $team_players['row_count']; $i++) {
   array_push($current_players, intval($team_players['rows_affected'][$i]['RLPID']));
 }
 
-// Get clicked player button
-foreach ($_POST as $button => $val) {
-  $button_info = explode('-', $button);
-  $RLPID = intval($button_info[1]);
-  switch ($button_info[2]) {
-    case 'add':
-      if (count($current_players) < 5) {
-        $timestamp = date('Y-m-d H:i:s');
-        $add_sql = "INSERT INTO TeamPlayer VALUES (?, ?, ?, 0, 0, 0)";
-        $addinfo = execute_query($add_sql, array($RLPID, $teamid, $timestamp));
-        array_push($current_players, $RLPID);
-      }
-      break;
-    case 'remove':
-      if (count($current_players) > 0) {
-        $remove_sql = "DELETE FROM TeamPlayer WHERE TeamID=? AND RLPID=?";
-        // Delete player from team in DB
-        $removeinfo = execute_query($remove_sql, array($teamid, $RLPID));
-        if (($idx = array_search($RLPID, $current_players)) !== false) {
-          unset($current_players[$idx]);
+
+$sql = "SELECT RLPID, name, position, mvps, goals, assists FROM RLPlayer";
+$search_params = array();
+
+if (isset($_POST['POST-TYPE'])) {
+  if ($_POST['POST-TYPE'] == 'Add/Remove') {
+    // Get clicked player button
+    foreach ($_POST as $button => $val) {
+      $button_info = explode('-', $button);
+      if (count($button_info) == 3) {
+        $RLPID = intval($button_info[1]);
+        switch ($button_info[2]) {
+          case 'add':
+            if (count($current_players) < 5) {
+              $timestamp = date('Y-m-d H:i:s');
+              $add_sql = "INSERT INTO TeamPlayer VALUES (?, ?, ?, 0, 0, 0)";
+              $addinfo = execute_query($add_sql, array($RLPID, $teamid, $timestamp));
+              array_push($current_players, $RLPID);
+            }
+            break;
+          case 'remove':
+            if (count($current_players) > 0) {
+              $remove_sql = "DELETE FROM TeamPlayer WHERE TeamID=? AND RLPID=?";
+              // Delete player from team in DB
+              $removeinfo = execute_query($remove_sql, array($teamid, $RLPID));
+              if (($idx = array_search($RLPID, $current_players)) !== false) {
+                unset($current_players[$idx]);
+              }
+            }
+            break;
+          default:
+            echo "ERROR";
+            break;
         }
       }
-      break;
-    default:
-      echo "ERROR";
-      break;
+    }
+  } else if ($_POST['POST-TYPE'] == 'Search') {
+    // Update results to match search terms
+    $searchInput = "%" . $_POST['player-search'] . "%";
+    $sql .= ' WHERE name LIKE ?;';
+    array_push($search_params, $searchInput);
   }
 }
+
+$data = execute_query($sql, $search_params);
 
 function actionsTD($RLPID)
 {
@@ -61,23 +74,23 @@ function actionsTD($RLPID)
             font-weight: 700;
             cursor: pointer;';
 
-    if (in_array($RLPID, $current_players)) {
-      $style .= ' background-color: red;';
-      $action = "Remove";
-      $RLPID .= '-remove';
-    } else {
-      if (count($current_players) == 5) {
-        $style .= ' background-color: gray;';
-        echo '<div class="ps-actions">
+  if (in_array($RLPID, $current_players)) {
+    $style .= ' background-color: red;';
+    $action = "Remove";
+    $RLPID .= '-remove';
+  } else {
+    if (count($current_players) == 5) {
+      $style .= ' background-color: gray;';
+      echo '<div class="ps-actions">
             <button type="submit" style="' . $style . ' cursor: unset;" disabled>Roster Full</button>
             </div>';
-        return;
-      }
-      $style .= ' background-color: green;';
-      $action = "Add";
-      $RLPID .= '-add';
+      return;
     }
-    echo '<div class="ps-actions">
+    $style .= ' background-color: green;';
+    $action = "Add";
+    $RLPID .= '-add';
+  }
+  echo '<div class="ps-actions">
           <button name="button-' . $RLPID . '" type="submit" style="' . $style . '">' . $action . ' Player</button>
         </div>';
 }
@@ -92,8 +105,9 @@ function textTD($contents)
 <link rel="stylesheet" href="css/player-search.css">
 <div class="inner-page-contents">
   <div class="player-search">
-    <div class="player-search-top-section">
+    <form class="player-search-top-section" action="player-search" method="post">
       <div class="ps-input">
+        <input type="text" name="POST-TYPE" value="Search" style="display: none;">
         <input type="text" name="player-search" id="player-search" placeholder="Search Player Info..." autofocus>
         <i class="fas fa-search"></i>
       </div>
@@ -108,9 +122,12 @@ function textTD($contents)
           <option value="MVPs">MVPs</option>
         </select>
       </div>
-    </div>
+      <input type="submit" style="display: none" />
+    </form>
+
     <div class="player-search-bottom-section">
       <form class="table-wrapper" action="player-search" method="post">
+        <input type="text" name="POST-TYPE" value="Add/Remove" style="display: none;">
         <table>
           <thead>
             <tr class="table-header">
@@ -123,7 +140,8 @@ function textTD($contents)
             </tr>
           </thead>
           <tbody>
-            <?php for ($i = 0; $i < $data['row_count']; $i++) :
+            <?php $i = 0;
+            for (; $i < $data['row_count']; $i++) :
               $player = $data['rows_affected'][$i]; ?>
               <tr>
                 <td><?php textTD($player['name']) ?></td>
@@ -133,10 +151,20 @@ function textTD($contents)
                 <td><?php textTD($player['mvps']) ?></td>
                 <td style="vertical-align: top;"><?php actionsTD($player['RLPID']) ?></td>
               </tr>
+            <?php endfor;
+            for (; $i < 5; $i++) : ?>
+            <tr>
+              <td></td>
+              <td></td>
+              <td></td>
+              <td></td>
+              <td></td>
+              <td></td>
+            </tr>
             <?php endfor; ?>
           </tbody>
         </table>
-        </form>
+      </form>
     </div>
   </div>
 </div>
